% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_security.R
\name{clear_session_state}
\alias{clear_session_state}
\title{Clear Session State}
\usage{
clear_session_state(state_reactive)
}
\arguments{
\item{state_reactive}{ReactiveVal containing session state}
}
\value{
Cleared state
}
\description{
Clears all authentication data from session.
}
\keyword{internal}
