% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Startup Message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\value{
NULL (invisible)
}
\description{
Display message when package is loaded.
}
\keyword{internal}
