% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supabase_client.R
\name{get_user_info}
\alias{get_user_info}
\title{Get user information}
\usage{
get_user_info(client, access_token)
}
\arguments{
\item{client}{A Supabase client created with \code{supabase_client()}}

\item{access_token}{User's access token}
}
\value{
User information or NULL if failed
}
\description{
Retrieve current user information using access token.
}
