% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_security.R
\name{refresh_access_token}
\alias{refresh_access_token}
\title{Refresh Access Token}
\usage{
refresh_access_token(client, refresh_token)
}
\arguments{
\item{client}{Supabase client object}

\item{refresh_token}{Refresh token}
}
\value{
List with new tokens or error
}
\description{
Uses refresh token to obtain a new access token from Supabase.
}
\keyword{internal}
