% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_modules.R
\name{supabase_logout_server}
\alias{supabase_logout_server}
\title{Logout Server Module}
\usage{
supabase_logout_server(id, client, user_state)
}
\arguments{
\item{id}{Module namespace ID}

\item{client}{Supabase client object}

\item{user_state}{Reactive containing current user state}
}
\value{
Updated user state reactive
}
\description{
Server logic for handling logout.
}
