% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supabase_client.R
\name{supabase_signout}
\alias{supabase_signout}
\title{Sign out user}
\usage{
supabase_signout(client, access_token)
}
\arguments{
\item{client}{A Supabase client created with \code{supabase_client()}}

\item{access_token}{User's current access token}
}
\value{
A list with success status
}
\description{
Sign out the current user and invalidate their session.
}
